requirements.txt
openai>=1.12.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
boto3